feat: add config set flag (#177)
* Adds support for ldcli config --set

The set flag takes a key/value pair and writes them to the config file
if they are relevant. It ignores arbitrary keys and for now only
supports:

* access-token
* base-uri

* Add config unset flag (#178)

The unset flag takes a value removes it from the config file if
it is relevant. It ignores arbitrary keys and for now only supports:

access-token
base-uri